@@ -15,7 +15,7 @@ import 'package:jordan/screens/prayers_screen.dart';
 import 'package:jordan/extras/statics.dart';
 
 class HomePage extends StatefulWidget {
-  const HomePage({Key key}) : super(key: key);
+  const HomePage({Key? key}) : super(key: key);
 
   @override
   _HomePageState createState() => _HomePageState();
@@ -35,7 +35,7 @@ class _HomePageState extends State<HomePage> {
       backgroundColor: AppColors.background,
       appBar: AppBar(
         backgroundColor: AppColors.foreground,
-        title: Text("Modlitewnik Salwatorianina"),
+        title: Text(AppStrings.appName),
       ),
       drawer: DrawerWidget(scaffoldKey: scaffoldKey),
       body: SingleChildScrollView(
